Created Map.py
This program, when run through Script.sh, sorts through the image information of any folder using exiftool, and extracts the GPS location from the metadata. Then using Folium, the image coordinates are placed on a map which is then opened automatically in a web browser. The commented code is to display the image that correlates with the mark on the map, but that is still a work in progress as I need to find a way to make sure the image name is paired with the GPS Cooridinates, which is harder than I thought it would be.